requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-jwt==4.0.0

# PDF generation
weasyprint==60.2
reportlab==4.0.7

# QR codes and barcodes
qrcode[pil]==7.4.2
python-barcode[images]==0.15.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
pandas==2.1.3
openpyxl==3.1.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Date and time
pytz==2023.3

# Image processing
Pillow==10.1.0

# Logging
loguru==0.7.2